Modulacion
Creé token.h, pila.h, lista.h y los llamo como librerías en el principal.
